worker/worker_requirements.txt
accelerate==0.31.0
aiohttp==3.9.5
aiofiles==23.2.1
arq==0.26.0
asyncpg==0.29.0
langchain-huggingface==0.0.3
ollama==0.2.1
pgvector==0.2.5
pydantic-settings==2.3.0
python-ffmpeg==2.0.12
SQLAlchemy==2.0.30
--find-links https://download.pytorch.org/whl/torch_stable.html
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
transformers==4.41.2
uvicorn==0.30.1
